Improve process to build l10n bundle without comments
